backend/requirements.txt
Django==4.2.14
pillow==10.4.0
django-map-location==0.9.2
